Using ExtractShortcutVT to extract position data from shortcut task.
Modified ExtractShortcutVT so it retruns a value. expday.m contains
the recording files for each rat, used to speed up workflow. The
testing_pos.m is where I ran the script that called the ExtractShortcutVT
function and plotted the results to see what it looks like.

ExtractShortcutVT.m saves a *-vt.mat file containing the pos_tsd
(x,y in data and tvec).