Linearizing shortcut experiment to "ideal" track.
Example usage is in linearize_shortcut.m. This code linearizes rat position to ideal track, with points defined in expkeys. It is based on the actual position within a certain distance from the "ideal" track, and the other actual positions not included in that analysis. Separates the position to "u," "shortcut," "novel" tracks. Those ideal locations as well as the time intervals in which they occur are returned.

This step in the analysis shows "ideal" rat position, and is used in the decoder scripts.