possibly major change to MultiRaster
cfg.openInAxes wasn't particularly clear in its purpose. It also seemed
to be a logical in cfg_def, but maybe an axes handle in cfg_in if user
wants. Working around cfg.openInAxes being either a logical or a handle
while trying to get MR to work with subplot became annoying. I did not
want to code around this so I just removed it entirely.