easy Tmaze color schemes with TmazeColors
for plotting the same colors in different scripts. easy is a relative
term. this function is intended to make switching schemes easy, and does
not require copy-pasting rgb values for 2 conditions for 4 rats across
different scripts when minds are changed. rather, add a new colormode
and the new rgb values inside of the function.