New functions ProcessConfig and CheckConfig
ProcessConfig adopts the name of the original ProcessConfig script. It
works the same way as ProcessConfig2 but has the additional capability
of rudimentary spell-checking and will inform the user of unrecognized
fields if cfg.verbose = 1. ProcessConfig2 should still exist for cases
where spellchecking is not desired but the field cfg.verbose is wanted
for other parts of the function (if that makes sense?).
CheckConfig is also included as a separate function in case someone can
find a use for it.