Warning! Edit to LoadSpikes
LoadSpikes field cfg.useClustersFile has been changed to cfg.getRatings
to be similar to cfg.getTTnumbers, both pertaining to usr fields of
similar name.

Additionally, LoadSpikes no longer errors if a clusters file is missing;
instead it makes a warning and replaces that rating with NaN so that the
user can keep working and find the missing clusters file later.

Also it has been modified to work for .tt files.

I've used it for a month and haven't noticed any issues with it.

The original LoadSpikes has been moved to the graveyard.