changes to MakeTC and DetectPlaceCells1D
possibly addresses a comment in DetectPlaceCells1d: "% firing rate --
could improve by having knowledge of whole-session"

if S has a usr 'rate' field, this is added to the tc inside of MakeTC to
be passed long to DetectPlaceCells1D. (i haven't used TuningCurves so i
didn't modify it to pass the usr rate field along)

DetectPlaceCells1D can now accept tc as a tc struct (as well as the
original tc matrix). if tc is a struct and has the usr.rate field, then
cells are excluded based on the usr data.