TSDtoIV2 resizes interval boundaries to the mean
new option cfg.ResizeAtMean
This is done in a number of papers for SWR detection, and it seems to
work quite well in cases with isolated SWRs (will attach a screen shot
for example), but will merge nearby events and result in long events
that contain more than a SWR (such as cases where there's mild ripple
frequency content surrounding a ripple).
By doing this, we ensure that the duration of detected events is not
dependent on the z-score threshold: higher thresholds result in shorter
durations than lower thresholds. But it does have its own problems like
merging and unreasonably long events in some cases.