sidekick now plots iv data as patch objects
When plotting multiple iv data simultaneously, it can be hard to
distinguish which lines are tstarts and which are tends when there are
many intervals nearby. Now cfg.patch plots a transparent patch object
between tstarts and tends so that it's obvious (thought not as pretty,
depending on the color choices). Screenshot included.

As with ducktrap, if you are plotting massive numbers of spikes (like tt
files with MultiRaster) and you then use patch objects, the plotting
speed increases substantially (not sure why). This is also true if you
use MulitRaster followed by sidekick: MultiRaster(cfg,Stt);
sidekick([],SWRiv);

Color specification has been updated to be more flexible.