new function ConvTSD
I have a few functions that convolve TSD. So rather than repeating steps
across functions I made a single function they can all use.
It can also make valleys deeper (for separation of closely neighbouring
ripples), and peaks taller (not sure what for, but might as well code it
in anyway since it's so similar to the valley step).