Set up code-python structure
This sets up the directory structure in the code-python folder,
and initializes it with some generic functions in the vdmlab
package (which is installable using setup.py), and my shortcut
analysis in the projects/emily_shortcut folder, which uses the
functions in the vdmlab package.

The vdmlab functions are documented. The documentation is built
with Sphinx.