Adding data path for new user's computername
Original code required user to edit checkTmazeReqs() AND getTmazeDataPath() to add their machine name. Instead, there is an existing function called getBaseFp(). Make sure to modify all code to use this ONE function so that the user only needs to make one change. Added an error message to getBaseFP that prompts the user to edit the function and add their machine name if their machine is not recognized.